+++ src/components/JsonWindow/style.css
